Move add resources button deeper down tree
