+++ apps/nextjs/src/app/problems/[id]/_components/right.tsx
@@ -1,27 +1,14 @@
 import type { RouterOutputs } from "@feprep/api";
-import type { User } from "@feprep/auth";
-import { ChatBubbleIcon, PlusIcon, VideoIcon } from "@feprep/ui";
-import { Button } from "@feprep/ui/button";
-import {
-  Sheet,
-  SheetContent,
-  SheetDescription,
-  SheetHeader,
-  SheetTitle,
-  SheetTrigger,
-} from "@feprep/ui/sheet";
+import { ChatBubbleIcon, VideoIcon } from "@feprep/ui";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@feprep/ui/tabs";
 
-import { AddResourceForm } from "./add-resource-form";
 import { DiscussionTab } from "./discussion-tab";
 import { ResourcesTab } from "./resources-tab";
 
 export async function Right({
   question,
-  user,
 }: {
   question: NonNullable<RouterOutputs["questions"]["byId"]>;
-  user: User | null;
 }) {
   return (
     <div>
@@ -37,10 +24,6 @@ export async function Right({
               Resources
             </TabsTrigger>
           </TabsList>
-          {/* if the user is a teacher, then show the button to add resources */}
-          {user?.type === "Teacher" && (
-            <AddResourceButton question={question} />
-          )}
         </div>
 
         <TabsContent value="discussion">
@@ -51,31 +34,5 @@ export async function Right({
         </TabsContent>
       </Tabs>
     </div>
-  );
-}
-
-function AddResourceButton({
-  question,
-}: {
-  question: NonNullable<RouterOutputs["questions"]["byId"]>;
-}) {
-  return (
-    <Sheet>
-      <SheetTrigger asChild>
-        <Button>
-          <PlusIcon className="mr-2 h-4 w-4" />
-          Add Resource
-        </Button>
-      </SheetTrigger>
-      <SheetContent className="w-[400px] sm:w-[540px]">
-        <SheetHeader>
-          <SheetTitle>Add Resource</SheetTitle>
-          <SheetDescription>
-            Add a link to a resource for this question.
-          </SheetDescription>
-        </SheetHeader>
-        <AddResourceForm question={question} />
-      </SheetContent>
-    </Sheet>
   );
 }
